test(abs): split the API-key tests so the regression fails behaviourally, not at build
The first cut of these tests referenced ABSModeAPIKey and ResolveAPIKey, so
reverting absauth.go made the package fail to COMPILE. That proves the API is
new; it does not prove the behaviour changed — the distinction that has shipped
real bugs in this repo before.

absauth_apikey_test.go now names no symbol introduced by this change, so with
absauth.go reverted it still compiles and fails on the actual regression:

  valid admin API key rejected by the ABS surface:
  got 401 body={"error":"authentication required"}

Tests that must name the new symbols live in absauth_apikey_mode_test.go.

Also corrects the inactive-owner expectation to 403: the credential is valid and
was accepted, it is the USER who is denied, which is what absCheckUserActive
already reports on the CF and JWT paths.
