+++ internal/server/middleware/absauth_apikey_test.go
@@ -12,7 +12,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/falkcorp/audiobook-organizer/internal/auth"
 	"github.com/falkcorp/audiobook-organizer/internal/database"
 	"github.com/falkcorp/audiobook-organizer/internal/server/absauth"
 	"github.com/falkcorp/audiobook-organizer/internal/oauth"
@@ -84,6 +83,11 @@ func (h *apiKeyHarness) get(raw string) *httptest.ResponseRecorder {
 // That is what made the ABS surface untestable with the one credential we have,
 // and why the author count and cold-cache latency could not be measured after
 // PR #2122 deployed.
+// It deliberately asserts only BEHAVIOUR — no reference to any symbol added by
+// this change — so that reverting absauth.go leaves this test COMPILING and
+// failing on the real 401-vs-200 regression, rather than failing to build. A test
+// that only fails to compile proves the API is new; it does not prove the
+// behaviour changed.
 func TestABSAPIKey_ReachesABSSurface(t *testing.T) {
 	h := newAPIKeyHarness(t)
 	h.store.addUser(&database.User{ID: "u-admin", Username: "admin", Status: "active"})
@@ -93,10 +97,11 @@ func TestABSAPIKey_ReachesABSSurface(t *testing.T) {
 	if w.Code != http.StatusOK {
 		t.Fatalf("valid admin API key rejected by the ABS surface: got %d body=%s", w.Code, w.Body.String())
 	}
-	if !strings.Contains(w.Body.String(), ABSModeAPIKey) {
-		t.Fatalf("expected mode %q in body, got %s", ABSModeAPIKey, w.Body.String())
+	if !strings.Contains(w.Body.String(), "u-admin") {
+		t.Fatalf("expected the key owner to be bound as the current user, got %s", w.Body.String())
 	}
 }
+
 
 // A revoked key must not become a back door just because it is pointed at ABS.
 func TestABSAPIKey_RevokedRejected(t *testing.T) {
@@ -164,40 +169,3 @@ func TestABSAPIKey_NonPrefixedTokenNotTreatedAsKey(t *testing.T) {
 	}
 }
 
-// 🔑 Scope narrowing. A key must not reach MORE through ABS than through /api/v1.
-// Without the intersectPermissions call in Bind, a read-only key would silently
-// become a full-privilege key the moment it was pointed at an ABS route.
-func TestABSAPIKey_ScopesNarrowPermissions(t *testing.T) {
-	h := newAPIKeyHarness(t)
-	h.store.addUser(&database.User{ID: "u1", Username: "admin", Status: "active", Roles: []string{"admin"}})
-	h.store.roles["admin"] = &database.Role{
-		ID:          "admin",
-		Permissions: []string{string(auth.PermLibraryView), string(auth.PermSettingsManage)},
-	}
-	// Scoped to VIEW only, though the owner's role also grants manage.
-	h.store.addKey("abk_scoped", &database.APIKey{
-		ID: "k1", UserID: "u1", Status: "active",
-		Scopes: []string{string(auth.PermLibraryView)},
-	})
-
-	c, _ := gin.CreateTestContext(httptest.NewRecorder())
-	c.Request = httptest.NewRequest(http.MethodGet, "/api/me", nil)
-	c.Request.Header.Set("Authorization", "Bearer abk_scoped")
-
-	id, aerr := h.resolver.ResolveAPIKey(c)
-	if aerr != nil || id == nil {
-		t.Fatalf("expected the scoped key to resolve, got id=%v err=%v", id, aerr)
-	}
-	h.resolver.Bind(c, id)
-
-	if !absHasPerm(c, auth.PermLibraryView) {
-		t.Fatal("scoped permission library:view was lost")
-	}
-	if absHasPerm(c, auth.PermSettingsManage) {
-		t.Fatal("a key scoped to library:view escalated to settings:manage on the ABS surface")
-	}
-}
-
-func absHasPerm(c *gin.Context, want auth.Permission) bool {
-	return auth.Can(c.Request.Context(), want)
-}
